backend/requirements.txt
# Core API framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
starlette>=0.27.0

# AI/ML dependencies
langchain>=0.3.0
langchain-core>=0.3.72
langsmith>=0.0.77
openai>=1.3.8
sentence-transformers>=2.2.2

# Vector database
pinecone>=5.0.0

# Configuration and utilities
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
requests>=2.32.3
packaging>=24.2

# Caching
redis>=5.0.1

# Development and testing
pytest>=7.4.3
tqdm>=4.66.1

# Rate limiting
slowapi>=0.1.9

# Google AI (Gemini)
langchain-google-genai>=1.0.1